+++ projects/_01_ingest/cqc_api/fargate/utils/locations_4_clean_utils.py
@@ -1,5 +1,6 @@
 import polars as pl
 
+from polars_utils import utils
 from utils.column_names.cleaned_data_files.cqc_location_cleaned import (
     CqcLocationCleanedColumns as CQCLClean,
 )
@@ -18,6 +19,35 @@ from utils.column_values.categorical_column_values import (
 from utils.value_labels.ind_cqc_filled_posts.primary_service_type_mapping import (
     CqcServiceToPrimaryServiceTypeSecondLevelLookup as PSSL_lookup,
 )
+
+
+def save_latest_full_snapshot(
+    cqc_lf: pl.LazyFrame, cqc_full_snapshot_destination: str
+) -> None:
+    """
+    Saves selected columns for the most recent import date as a parquet file.
+
+    All registered and deregistered locations in the latest full snapshot of CQC locations are saved.
+    This data is used in the Skills for Care (SfC) internal reconciliation process.
+
+    Args:
+        cqc_lf (pl.LazyFrame): LazyFrame containing CQC location data.
+        cqc_full_snapshot_destination (str): S3 URI to save CQC deregistered locations data.
+    """
+    required_columns = [
+        CQCLClean.cqc_location_import_date,
+        CQCLClean.location_id,
+        CQCLClean.registration_status,
+        CQCLClean.deregistration_date,
+    ]
+
+    cqc_lf = cqc_lf.select(*required_columns)
+
+    cqc_lf = utils.filter_to_maximum_value_in_column(
+        cqc_lf, CQCLClean.cqc_location_import_date
+    )
+
+    utils.sink_to_parquet(cqc_lf, cqc_full_snapshot_destination, append=False)
 
 
 def clean_provider_id_column(cqc_lf: pl.LazyFrame) -> pl.LazyFrame: