+++ utils/raw_data_adjustments.py
@@ -8,9 +8,6 @@ from utils.column_names.cleaned_data_files.ascwds_worker_cleaned import (
 from utils.column_names.cleaned_data_files.ascwds_workplace_cleaned import (
     AscwdsWorkplaceCleanedColumns as AWPClean,
 )
-from utils.column_names.cleaned_data_files.cqc_location_cleaned import (
-    NewCqcLocationApiColumns as CQCL,
-)
 
 
 def remove_duplicate_worker_in_raw_worker_data(raw_worker_df: DataFrame) -> DataFrame:
@@ -71,45 +68,3 @@ def remove_duplicate_workplaces_in_raw_workplace_data(
     )
 
     return raw_workplace_df
-
-
-# converted to polars -> polars_utils.raw_data_adjustments.py
-def remove_records_from_locations_data(
-    raw_locations_df: DataFrame,
-) -> DataFrame:
-    """
-    This function removes records from the locations dataset.
-    """
-    raw_locations_df = raw_locations_df.where(
-        (
-            raw_locations_df[CQCL.location_id]
-            != RecordsToRemoveInLocationsData.dental_practice
-        )
-        & (
-            raw_locations_df[CQCL.location_id]
-            != RecordsToRemoveInLocationsData.temp_registration
-        )
-    )
-    return raw_locations_df
-
-
-@dataclass
-class RecordsToRemoveInLocationsData:
-    """
-    This class contains the locations ids that should be removed from
-    the locations data.
-
-    Dental Practice:
-    The location is listed once as a social care org in the locations
-    dataset but is lited as Primary Dental Care on every other row and
-    in the providers dataset. The location ID is enough to identify
-    and remove this row.
-
-    Temporary Registration:
-    The location is listed once as registered in the locations dataset,
-    but conatins barely any data and appears to have deregistered very
-    quickly. The location ID is enough to identify and remove this row.
-    """
-
-    dental_practice: str = "1-12082335777"
-    temp_registration: str = "1-127367030"